refactor(mcp): extract deploy plan helper module
