+++ openspec/changes/refactor-mcp-deploy-plan-helper-module/README.md
@@ -1,0 +1,3 @@
+# refactor-mcp-deploy-plan-helper-module
+
+Refactor: extract MCP deploy-plan helper into a dedicated module
