+++ openspec/changes/refactor-mcp-deploy-plan-helper-module/tasks.md
@@ -1,0 +1,7 @@
+## 1. Implementation
+- [x] Extract `deploy_plan_envelope_in_process` into `src/mcp/tools/deploy_plan.rs`
+- [x] Keep MCP tool behavior and JSON envelopes unchanged
+
+## 2. Verification
+- [x] `cargo fmt --all`
+- [x] `just check`
