+++ src/mcp/tools.rs
@@ -13,6 +13,7 @@ use super::{AgentpackMcp, confirm};
 
 mod deploy;
 mod deploy_apply;
+mod deploy_plan;
 mod doctor;
 mod evolve_propose;
 mod evolve_restore;
@@ -20,6 +21,8 @@ mod explain;
 mod preview;
 mod rollback;
 mod status;
+
+use deploy_plan::deploy_plan_envelope_in_process;
 
 pub(super) const TOOLS_INSTRUCTIONS: &str = "Agentpack MCP server (stdio). Tools: plan, diff, preview, status, doctor, deploy, deploy_apply, rollback, evolve_propose, evolve_restore, explain.";
 
@@ -276,50 +279,6 @@ async fn call_evolve_propose_in_process(
 
 async fn call_preview_in_process(args: PreviewArgs) -> anyhow::Result<(String, serde_json::Value)> {
     preview::call_preview_in_process(args).await
-}
-
-async fn deploy_plan_envelope_in_process(args: CommonArgs) -> anyhow::Result<serde_json::Value> {
-    tokio::task::spawn_blocking(move || {
-        let repo_override = args.repo.as_ref().map(std::path::PathBuf::from);
-        let profile = args.profile.as_deref().unwrap_or("default");
-        let target = args.target.as_deref().unwrap_or("all");
-        let machine_override = args.machine.as_deref();
-
-        let result = crate::handlers::read_only::read_only_context(
-            repo_override.as_deref(),
-            machine_override,
-            profile,
-            target,
-        );
-
-        match result {
-            Ok(crate::handlers::read_only::ReadOnlyContext {
-                targets,
-                plan,
-                warnings,
-                ..
-            }) => {
-                let data =
-                    crate::app::deploy_json::deploy_json_data_dry_run(profile, targets, plan);
-                let mut envelope = crate::output::JsonEnvelope::ok("deploy", data);
-                envelope.warnings = warnings;
-                serde_json::to_value(&envelope).context("serialize deploy envelope")
-            }
-            Err(err) => {
-                let user_err = err.chain().find_map(|e| e.downcast_ref::<UserError>());
-                let code = user_err
-                    .map(|e| e.code.clone())
-                    .unwrap_or_else(|| "E_UNEXPECTED".to_string());
-                let message = user_err
-                    .map(|e| e.message.clone())
-                    .unwrap_or_else(|| err.to_string());
-                let details = user_err.and_then(|e| e.details.clone());
-                Ok(envelope_error("deploy", &code, &message, details))
-            }
-        }
-    })
-    .await
-    .context("mcp deploy handler task join")?
 }
 
 fn tool_result_from_envelope(text: String, envelope: serde_json::Value) -> CallToolResult {
